requirements.txt
PyOpenGL
PyOpenGL_accelerate
moderngl
tree-sitter
PyGLM
numpy
opencv-python
maxminddb-geolite2
wget
requests
chardet
build
setuptools
twine
wheel
pybind11
pyflakes
pyinstaller